requirements.txt
numpy
matplotlib
Pillow
tifffile
xraylib
